requirements.txt
Django~=3.2.20
pytest~=7.4.0
setuptools~=68.0.0
django-model-utils~=4.3.1

djangorestframework~=3.14.0
